requirements.txt
google-adk[database]==0.3.0
google-generativeai==0.8.5
psutil==5.9.5
python-dotenv==1.1.0
deprecated
pydantic